app/ippon_assistant/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "p1"
version = "0.1.0"
description = "AgentCore Runtime Application using Strands SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aws-opentelemetry-distro",
    "bedrock-agentcore >= 1.0.3",
    "botocore[crt] >= 1.35.0",
    "mcp >= 1.19.0",
    "strands-agents >= 1.13.0",
    "fastapi >= 0.115.0",
    "uvicorn >= 0.30.0",
    "python-multipart >= 0.0.9",
    "httpx >= 0.27.0",
    "mangum >= 0.17.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
